+++ config/chai.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = {
+	globals: {
+		expect: true,
+	},
+	extends: [
+		'../rules/plugins/mocha',
+	].map(require.resolve),
+	rules: {
+		// chai uses what appears to be unused expressions
+		'no-unused-expressions': 'off',
+	},
+};