+++ config/ecmascript-7.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+	parserOptions: {
+		ecmaVersion: 7,
+	},
+	extends: [
+		'./ecmascript-6',
+	].map(require.resolve),
+};