+++ rules/custom.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-	plugins: [ 'strict-newline' ],
-	rules: { 'strict-newline/strict-newline': 'error' }
-};