+++ rules/ecmascript-8.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+	rules: {
+		'promise/prefer-await-to-then': 'error',
+		'promise/prefer-await-to-callbacks': 'error',
+	},
+};