+++ rules/plugins/babel.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = {
+	rules: {
+		'new-cap': 'off',
+		'object-curly-spacing': 'off',
+		'no-invalid-this': 'off',
+		semi: 'off',
+		'babel/new-cap': 'error',
+		'babel/object-curly-spacing': 'error',
+		'babel/no-invalid-this': 'error',
+		'babel/semi': 'error',
+	},
+};