+++ rules/plugins/promise.js
@@ -1,0 +1,16 @@
+'use strict';
+
+module.exports = {
+	rules: {
+		'promise/always-return': 'error',
+		'promise/no-return-wrap': 'error',
+		'promise/param-names': 'error',
+		'promise/catch-or-return': 'error',
+		'promise/no-native': 'error',
+		'promise/no-nesting': 'off',
+		'promise/no-promise-in-callback': 'off',
+		'promise/no-callback-in-promise': 'off',
+		'promise/avoid-new': 'off',
+		'promise/no-return-in-finally': 'error',
+	},
+};