+++ rules/strict-mode.js
@@ -1,8 +1,7 @@
-/* eslint object-curly-newline: off */
 'use strict';
 
 module.exports = {
 	rules: {
-		strict: ['error', 'global']
-	}
+		strict: 'error',
+	},
 };