+++ CHANGELOG.md
@@ -3,6 +3,21 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 The format conforms to [Keep a Changelog](http://keepachangelog.com/).
 
+## [3.0.0] - 2017-12-15
+### Added
+- ECMAScript 6 support
+- ECMAScript 7 support
+- ECMAScript 8 support
+- [Babel](https://github.com/babel/eslint-plugin-babel) plugin
+- [Filenames](https://github.com/selaux/eslint-plugin-filenames) plugin
+- [Mocha](https://github.com/lo1tuma/eslint-plugin-mocha) plugin
+- [Promise](https://github.com/xjamundx/eslint-plugin-promise) plugin
+- [React](https://github.com/yannickcr/eslint-plugin-react) plugin
+- [Security](https://github.com/nodesecurity/eslint-plugin-security) plugin
+
+### Changed
+- Updated to support ESLint 4
+
 ## [2.0.0] - 2016-07-24
 ### Changed
 - Updated to support ESLint 3